+++ config/laravel_backup_panel.php
@@ -29,6 +29,6 @@ return [
     |
     */
 
-    'queue' => null,
+    'queue' => 'default',
 
 ];