+++ resources/views/layout.blade.php
@@ -4,24 +4,15 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <title>Laravel Backup Panel{{ config('app.name') ? ' - ' . config('app.name') : '' }}</title>
+    <title>Laravel Backup Panel</title>
 
-    <!-- Fonts -->
-    <link rel="dns-prefetch" href="//fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css?family=Nunito" rel="stylesheet">
-
-    <!-- Styles -->
-    <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/toastify-js/src/toastify.min.css">
-    <link href="{{ asset('vendor/laravel_backup_panel/bootstrap.css') }}" rel="stylesheet">
-    <link href="{{ asset('vendor/laravel_backup_panel/app.css') }}" rel="stylesheet">
-    @livewireStyles
+    <link href="{{ asset('vendor/laravel_backup_panel/css/bootstrap.min.css') }}" rel="stylesheet">
+    <link href="{{ asset('vendor/laravel_backup_panel/css/app.css') }}" rel="stylesheet">
 </head>
 <body>
-    <livewire:laravel_backup_panel::app />
+    @yield('content')
 
-    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ho+j7jyWK8fNQe+A12Hb8AhRq26LrZ/JpcUGGOn+Y7RsweNrtN/tE3MoK7ZeZDyx" crossorigin="anonymous"></script>
-    <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/toastify-js"></script>
-    @livewireScripts
+    <script src="{{ asset('vendor/laravel_backup_panel/js/bootstrap.bundle.min.js') }}"></script>
+    <script src="{{ asset('vendor/laravel_backup_panel/js/app.js') }}"></script>
 </body>
 </html>