+++ src/LaravelBackupPanelServiceProvider.php
@@ -2,11 +2,10 @@
 
 namespace PavelMironchik\LaravelBackupPanel;
 
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
-use Livewire\Livewire;
 use PavelMironchik\LaravelBackupPanel\Console\InstallCommand;
-use PavelMironchik\LaravelBackupPanel\Http\Livewire\App;
 use PavelMironchik\LaravelBackupPanel\Http\Middleware\Authenticate;
 
 class LaravelBackupPanelServiceProvider extends ServiceProvider
@@ -16,7 +15,7 @@ class LaravelBackupPanelServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function boot()
+    public function boot(): void
     {
         if ($this->app->runningInConsole()) {
             $this->publishes([
@@ -24,7 +23,7 @@ class LaravelBackupPanelServiceProvider extends ServiceProvider
             ], 'laravel-backup-panel-config');
 
             $this->publishes([
-                __DIR__.'/../resources/assets/css' => public_path('vendor/laravel_backup_panel'),
+                __DIR__.'/../resources/assets' => public_path('vendor/laravel_backup_panel'),
             ], 'laravel-backup-panel-assets');
 
             $this->publishes([
@@ -41,15 +40,13 @@ class LaravelBackupPanelServiceProvider extends ServiceProvider
         }
 
         Route::group([
-            'prefix' => config('laravel_backup_panel.path'),
+            'prefix' => Config::string('laravel_backup_panel.path'),
             'middleware' => ['web', Authenticate::class],
         ], function () {
             $this->loadRoutesFrom(__DIR__.'/../routes/web.php');
         });
 
         $this->loadViewsFrom(__DIR__.'/../resources/views', 'laravel_backup_panel');
-
-        Livewire::component('laravel_backup_panel::app', App::class);
     }
 
     /**
@@ -57,7 +54,7 @@ class LaravelBackupPanelServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function register()
+    public function register(): void
     {
         $this->mergeConfigFrom(__DIR__.'/../config/laravel_backup_panel.php', 'laravel_backup_panel');
     }