+++ src/Rules/BackupDisk.php
@@ -3,17 +3,16 @@
 namespace PavelMironchik\LaravelBackupPanel\Rules;
 
 use Illuminate\Contracts\Validation\Rule;
+use Spatie\Backup\Config\Config as BackupConfig;
 
 class BackupDisk implements Rule
 {
-    public function passes($attribute, $value)
+    public function passes($attribute, $value): bool
     {
-        $configuredBackupDisks = config('backup.backup.destination.disks');
-
-        return in_array($value, $configuredBackupDisks);
+        return in_array($value, app(BackupConfig::class)->backup->destination->disks, true);
     }
 
-    public function message()
+    public function message(): string
     {
         return 'Current disk is not configured as a backup disk';
     }